# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Stroit Backend Suite API
  version: 1.0.0
  description: |
    API surface for the Stroit project gateway and microservices.
    Paths are versioned (`/v1`) and all responses follow the `{ success, data, error }` contract.
servers:
  - url: https://dev.api.stroitproject.local
    description: Development environment
  - url: https://test.api.stroitproject.local
    description: Test environment
  - url: https://api.stroitproject.local
    description: Production environment
...
paths:
...
  /v1/users/register:
    post:
      tags:
        - Users
      summary: Register a new consumer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: Created user identifier
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/responses/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [id, email]
                        properties:
                          id:
                            type: string
                            format: uuid
                          email:
                            type: string
        "409":
          $ref: "#/components/responses/Conflict"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/users/login:
    post:
      tags:
        - Users
      summary: Authenticate and receive a Bearer JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/responses/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                          expiresIn:
                            type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/users/me:
    get:
      tags:
        - Users
      summary: Return the caller's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

    put:
      tags:
        - Users
      summary: Update the caller's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          $ref: "#/components/responses/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/users:
    get:
      tags:
        - Users
      summary: Paginated list of users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: perPage
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - in: query
          name: sort
          schema:
            type: string
            enum:
              - createdAt
              - email
              - name
            default: createdAt
        - in: query
          name: order
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
        - in: query
          name: role
          schema:
            type: string
            enum:
              - admin
              - user
        - in: query
          name: email
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/UserList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserBase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserRegistration:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    UserProfileUpdate:
      type: object
      required: []
      properties:
        name:
          type: string
        password:
          type: string
          minLength: 8
...
    UserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserBase"
        page:
          type: integer
        perPage:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    ErrorPayload:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                const: true
              data:
                type: object
    User:
      description: User data
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/responses/Success"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/UserBase"
...
    UserList:
      description: Paginated user list
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/responses/Success"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/UserList"
...
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                const: false
              error:
                $ref: "#/components/schemas/ErrorPayload"
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                const: false
              error:
                $ref: "#/components/schemas/ErrorPayload"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                const: false
              error:
                $ref: "#/components/schemas/ErrorPayload"
...
    Conflict:
      description: Entity conflict (e.g., duplicate email)
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                const: false
              error:
                $ref: "#/components/schemas/ErrorPayload"